+++ stencil.config.ts
@@ -26,6 +26,7 @@ export const config: Config = {
       ]
     },
     { components: ["calcite-icon"] },
+    { components: ["calcite-link"] },
     { components: ["calcite-loader"] },
     { components: ["calcite-modal"] },
     { components: ["calcite-notice"] },